requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
psycopg2-binary>=2.9.9
Werkzeug==3.0.1
PyPDF2==3.0.1
Pillow>=10.2.0
python-dotenv==1.0.0
requests>=2.31.0
Flask-SocketIO==5.3.6
simple-websocket==1.0.0
google-generativeai>=0.8.3
gunicorn==21.2.0

